+++ Source/Projection.h
@@ -13,8 +13,8 @@ public:
 	QProjection::QProjection(const QProjection& Other);
 	QProjection& QProjection::operator=(const QProjection& Other);
 
-	int				GetFieldOfView(void) const;
-	void			SetFieldOfView(const int& FieldOfView);
+	float			GetFieldOfView(void) const;
+	void			SetFieldOfView(const float& FieldOfView);
 	void			Reset(void);
 	void			ReadXML(QDomElement& Parent);
 	QDomElement		WriteXML(QDomDocument& DOM, QDomElement& Parent);
